// frontend/src/types/dto.ts
export type Fish = {
    id: number
    project: string
    trip: string
    longitude: number
    latitude: number
    dateTime: Date
    station: string
    bottDepthM: number
    sampleId: number
    parameter: string
    tissue: string
    species: string
    individuals: number
    value: number
    units: string
    quality: string
}